refactor uints to ints everywhere

+++ gif/animator.go
@@ -62,8 +62,8 @@ func (a *Animator) Encode(w io.Writer) error {
 
 func (a *Animator) gif() *gif.GIF {
 	first := a.addedGrids[0]
-	widthPixels := int(first.Width()) * a.resolution
-	heightPixels := int(first.Height()) * a.resolution
+	widthPixels := first.Width() * a.resolution
+	heightPixels := first.Height() * a.resolution
 	return &gif.GIF{
 		Image:     a.images(),
 		Delay:     a.delay(),
@@ -89,8 +89,8 @@ func (a *Animator) images() []*image.Paletted {
 
 // GridToImage transforms a grid into an image.
 func (a *Animator) gridToImage(g conway.Grid) *image.Paletted {
-	width := int(g.Width()) * a.resolution
-	height := int(g.Height()) * a.resolution
+	width := g.Width() * a.resolution
+	height := g.Height() * a.resolution
 	r := image.Rectangle{
 		Min: image.Point{
 			X: 0,
@@ -120,7 +120,7 @@ func (a *Animator) gridToImage(g conway.Grid) *image.Paletted {
 	const black = 1
 	for c := 0; c < width; c++ {
 		for r := 0; r < height; r++ {
-			cell := coord.New(uint(c/a.resolution), uint(r/a.resolution))
+			cell := coord.New(c/a.resolution, r/a.resolution)
 			isAlive, err := g.IsAlive(cell)
 			if err != nil {
 				panic(err) // unreachable
